feat: Add `delete tx` tests and improve existing

+++ src/controllers/transactions.controller/delete-transaction.ts
@@ -1,41 +1,24 @@
-import { API_ERROR_CODES, API_RESPONSE_STATUS } from 'shared-types';
+import { API_RESPONSE_STATUS } from 'shared-types';
 import { CustomResponse } from '@common/types';
-
-import { CustomError } from '@js/errors'
+import { errorHandler } from '@controllers/helpers';
 
 import * as transactionsService from '@services/transactions';
+import { ValidationError } from '@js/errors';
 
 export const deleteTransaction = async (req, res: CustomResponse) => {
   try {
-    const { id, authorId } = req.params;
+    const { id } = req.params;
     const { id: userId } = req.user;
 
-    await transactionsService.deleteTransaction({
-      id,
-      userId: authorId ?? userId,
-    })
+    if (!id || !Number(id)) throw new ValidationError({ message: 'Transaction id parameter does not specified in the URL' });
+
+    await transactionsService.deleteTransaction({ id, userId })
 
     return res.status(200).json({
       status: API_RESPONSE_STATUS.success,
       response: {},
     });
   } catch (err) {
-    if (err instanceof CustomError) {
-      return res.status(err.httpCode).json({
-        status: API_RESPONSE_STATUS.error,
-        response: {
-          message: err.message,
-          code: err.code,
-        },
-      });
-    }
-
-    return res.status(500).json({
-      status: API_RESPONSE_STATUS.error,
-      response: {
-        message: 'Unexpected error.',
-        code: API_ERROR_CODES.unexpected,
-      },
-    });
+    errorHandler(res, err);
   }
 };
